pyproject.toml
[project]
name = "webex-room-id-fetcher"
version = "0.1.0"
description = "A command-line tool to quickly find Webex room/space IDs by name using the Webex API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.0",
    "rich>=14.1.0",
    "typer>=0.16.0",
    "webexpythonsdk>=2.0.5",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.9",
    "xenon>=0.9.3",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "W",  # pycodestyle warnings
    "C90",# flake8-mccabe complexity
    "N",  # pep8-naming
    "UP", # pyupgrade
    "ANN",# flake8-annotations
    "D",  # pydocstyle
    "T",  # flake8-print (includes T201 for pdb, replacing debug-statements hook)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
combine-as-imports = true
